api/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.3.7",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "groq>=0.37.1",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "supabase>=2.27.0",
]
